Add: Seperate multiple image and single image controller for add book and update
